moif/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moif"
version = "0.1.0"
description = "Meaning-Oriented Invariance Framework (MOIF) v0.1 - reproducible OSS research engine for invariance breaking"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Kotaro Furukawa" }]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.10",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "jsonschema>=4.20",
  "jinja2>=3.1"
]

[project.scripts]
moif = "moif.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["moif*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E","F","I","UP","B"]
